# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Settings {
  id: number;
  dailyGoal: number;
}

export interface UpdateSettingsBody {
  /**
   * @minimum 500
   * @maximum 10000
   */
  dailyGoal: number;
}

export type FoodEntryMeal = (typeof FoodEntryMeal)[keyof typeof FoodEntryMeal];

export const FoodEntryMeal = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface FoodEntry {
  id: number;
  name: string;
  calories: number;
  quantity: string;
  meal: FoodEntryMeal;
  date: string;
  createdAt: string;
}

export type CreateFoodEntryBodyMeal =
  (typeof CreateFoodEntryBodyMeal)[keyof typeof CreateFoodEntryBodyMeal];

export const CreateFoodEntryBodyMeal = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface CreateFoodEntryBody {
  name: string;
  /** @minimum 0 */
  calories: number;
  quantity: string;
  meal: CreateFoodEntryBodyMeal;
  date?: string;
}

export type DailySummaryBreakdownByMeal = {
  breakfast: number;
  lunch: number;
  dinner: number;
  snack: number;
};

export type DailySummaryStatus =
  (typeof DailySummaryStatus)[keyof typeof DailySummaryStatus];

export const DailySummaryStatus = {
  on_track: "on_track",
  warning: "warning",
  over_goal: "over_goal",
} as const;

export interface DailySummary {
  date: string;
  totalCalories: number;
  dailyGoal: number;
  remaining: number;
  percentUsed: number;
  breakdownByMeal: DailySummaryBreakdownByMeal;
  status: DailySummaryStatus;
}

export interface DayStats {
  date: string;
  totalCalories: number;
  dailyGoal: number;
}

export interface Recipe {
  id: number;
  title: string;
  description: string;
  ingredients: string;
  instructions: string;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  photoPath?: string | null;
  isFavorite: boolean;
  createdAt: string;
}

export interface CreateRecipeBody {
  title: string;
  description?: string;
  ingredients: string;
  instructions: string;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  photoPath?: string | null;
}

export interface ToggleFavoriteBody {
  isFavorite: boolean;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListFoodEntriesParams = {
  date?: string;
};

export type GetDailySummaryParams = {
  date?: string;
};

export type ListRecipesParams = {
  favoritesOnly?: string;
};