# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Calorie Tracker API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: settings
    description: User settings (daily calorie goal)
  - name: food-entries
    description: Food diary entries
  - name: summary
    description: Daily summary and stats
  - name: recipes
    description: User recipes
  - name: storage
    description: Object storage
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get user settings
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

  /food-entries:
    get:
      operationId: listFoodEntries
      tags: [food-entries]
      summary: List food entries for a given date
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of food entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodEntry"
    post:
      operationId: createFoodEntry
      tags: [food-entries]
      summary: Add a food entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFoodEntryBody"
      responses:
        "201":
          description: Created food entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodEntry"

  /food-entries/{id}:
    delete:
      operationId: deleteFoodEntry
      tags: [food-entries]
      summary: Delete a food entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /summary:
    get:
      operationId: getDailySummary
      tags: [summary]
      summary: Get daily calorie summary
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Daily summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummary"

  /weekly-stats:
    get:
      operationId: getWeeklyStats
      tags: [summary]
      summary: Get calorie totals for the last 7 days
      responses:
        "200":
          description: Weekly stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DayStats"

  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List all recipes
      parameters:
        - name: favoritesOnly
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeBody"
      responses:
        "201":
          description: Created recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"

  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /recipes/{id}/favorite:
    patch:
      operationId: toggleRecipeFavorite
      tags: [recipes]
      summary: Toggle recipe favorite status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleFavoriteBody"
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Settings:
      type: object
      properties:
        id:
          type: integer
        dailyGoal:
          type: integer
      required: [id, dailyGoal]

    UpdateSettingsBody:
      type: object
      properties:
        dailyGoal:
          type: integer
          minimum: 500
          maximum: 10000
      required: [dailyGoal]

    FoodEntry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        calories:
          type: integer
        quantity:
          type: string
        meal:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        date:
          type: string
        createdAt:
          type: string
      required: [id, name, calories, quantity, meal, date, createdAt]

    CreateFoodEntryBody:
      type: object
      properties:
        name:
          type: string
        calories:
          type: integer
          minimum: 0
        quantity:
          type: string
        meal:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        date:
          type: string
      required: [name, calories, quantity, meal]

    DailySummary:
      type: object
      properties:
        date:
          type: string
        totalCalories:
          type: integer
        dailyGoal:
          type: integer
        remaining:
          type: integer
        percentUsed:
          type: number
        breakdownByMeal:
          type: object
          properties:
            breakfast:
              type: integer
            lunch:
              type: integer
            dinner:
              type: integer
            snack:
              type: integer
          required: [breakfast, lunch, dinner, snack]
        status:
          type: string
          enum: [on_track, warning, over_goal]
      required: [date, totalCalories, dailyGoal, remaining, percentUsed, breakdownByMeal, status]

    DayStats:
      type: object
      properties:
        date:
          type: string
        totalCalories:
          type: integer
        dailyGoal:
          type: integer
      required: [date, totalCalories, dailyGoal]

    Recipe:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        ingredients:
          type: string
        instructions:
          type: string
        calories:
          type: ["integer", "null"]
        photoPath:
          type: ["string", "null"]
        isFavorite:
          type: boolean
        createdAt:
          type: string
      required: [id, title, description, ingredients, instructions, isFavorite, createdAt]

    CreateRecipeBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ingredients:
          type: string
        instructions:
          type: string
        calories:
          type: ["integer", "null"]
        photoPath:
          type: ["string", "null"]
      required: [title, ingredients, instructions]

    ToggleFavoriteBody:
      type: object
      properties:
        isFavorite:
          type: boolean
      required: [isFavorite]

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required: [name, size, contentType]

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required: [uploadURL, objectPath]