# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get user settings
 */
export const GetSettingsResponse = zod.object({
  id: zod.number(),
  dailyGoal: zod.number(),
});

/**
 * @summary Update user settings
 */
export const updateSettingsBodyDailyGoalMin = 500;
export const updateSettingsBodyDailyGoalMax = 10000;

export const UpdateSettingsBody = zod.object({
  dailyGoal: zod
    .number()
    .min(updateSettingsBodyDailyGoalMin)
    .max(updateSettingsBodyDailyGoalMax),
});

export const UpdateSettingsResponse = zod.object({
  id: zod.number(),
  dailyGoal: zod.number(),
});

/**
 * @summary List food entries for a given date
 */
export const ListFoodEntriesQueryParams = zod.object({
  date: zod.coerce.string().optional(),
});

export const ListFoodEntriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  calories: zod.number(),
  quantity: zod.string(),
  meal: zod.enum(["breakfast", "lunch", "dinner", "snack"]),
  date: zod.string(),
  createdAt: zod.string(),
});
export const ListFoodEntriesResponse = zod.array(ListFoodEntriesResponseItem);

/**
 * @summary Add a food entry
 */
export const createFoodEntryBodyCaloriesMin = 0;

export const CreateFoodEntryBody = zod.object({
  name: zod.string(),
  calories: zod.number().min(createFoodEntryBodyCaloriesMin),
  quantity: zod.string(),
  meal: zod.enum(["breakfast", "lunch", "dinner", "snack"]),
  date: zod.string().optional(),
});

/**
 * @summary Delete a food entry
 */
export const DeleteFoodEntryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get daily calorie summary
 */
export const GetDailySummaryQueryParams = zod.object({
  date: zod.coerce.string().optional(),
});

export const GetDailySummaryResponse = zod.object({
  date: zod.string(),
  totalCalories: zod.number(),
  dailyGoal: zod.number(),
  remaining: zod.number(),
  percentUsed: zod.number(),
  breakdownByMeal: zod.object({
    breakfast: zod.number(),
    lunch: zod.number(),
    dinner: zod.number(),
    snack: zod.number(),
  }),
  status: zod.enum(["on_track", "warning", "over_goal"]),
});

/**
 * @summary Get calorie totals for the last 7 days
 */
export const GetWeeklyStatsResponseItem = zod.object({
  date: zod.string(),
  totalCalories: zod.number(),
  dailyGoal: zod.number(),
});
export const GetWeeklyStatsResponse = zod.array(GetWeeklyStatsResponseItem);

/**
 * @summary List all recipes
 */
export const ListRecipesQueryParams = zod.object({
  favoritesOnly: zod.coerce.string().optional(),
});

export const ListRecipesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  ingredients: zod.string(),
  instructions: zod.string(),
  calories: zod.number().nullish(),
  photoPath: zod.string().nullish(),
  isFavorite: zod.boolean(),
  createdAt: zod.string(),
});
export const ListRecipesResponse = zod.array(ListRecipesResponseItem);

/**
 * @summary Create a recipe
 */
export const CreateRecipeBody = zod.object({
  title: zod.string(),
  description: zod.string().optional(),
  ingredients: zod.string(),
  instructions: zod.string(),
  calories: zod.number().nullish(),
  photoPath: zod.string().nullish(),
});

/**
 * @summary Get a recipe
 */
export const GetRecipeParams = zod.object({
  id: zod.coerce.number(),
});

export const GetRecipeResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  ingredients: zod.string(),
  instructions: zod.string(),
  calories: zod.number().nullish(),
  photoPath: zod.string().nullish(),
  isFavorite: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a recipe
 */
export const DeleteRecipeParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Toggle recipe favorite status
 */
export const ToggleRecipeFavoriteParams = zod.object({
  id: zod.coerce.number(),
});

export const ToggleRecipeFavoriteBody = zod.object({
  isFavorite: zod.boolean(),
});

export const ToggleRecipeFavoriteResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  ingredients: zod.string(),
  instructions: zod.string(),
  calories: zod.number().nullish(),
  photoPath: zod.string().nullish(),
  isFavorite: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Request a presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
});