# lib/api-zod/src/generated/types/createFoodEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFoodEntryBodyMeal } from "./createFoodEntryBodyMeal";

export interface CreateFoodEntryBody {
  name: string;
  /** @minimum 0 */
  calories: number;
  quantity: string;
  meal: CreateFoodEntryBodyMeal;
  date?: string;
}