# lib/api-zod/src/generated/types/createRecipeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRecipeBody {
  title: string;
  description?: string;
  ingredients: string;
  instructions: string;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  photoPath?: string | null;
}