# lib/api-zod/src/generated/types/dailySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DailySummaryBreakdownByMeal } from "./dailySummaryBreakdownByMeal";
import type { DailySummaryStatus } from "./dailySummaryStatus";

export interface DailySummary {
  date: string;
  totalCalories: number;
  dailyGoal: number;
  remaining: number;
  percentUsed: number;
  breakdownByMeal: DailySummaryBreakdownByMeal;
  status: DailySummaryStatus;
}