# lib/api-zod/src/generated/types/foodEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FoodEntryMeal } from "./foodEntryMeal";

export interface FoodEntry {
  id: number;
  name: string;
  calories: number;
  quantity: string;
  meal: FoodEntryMeal;
  date: string;
  createdAt: string;
}