# lib/api-zod/src/generated/types/recipe.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Recipe {
  id: number;
  title: string;
  description: string;
  ingredients: string;
  instructions: string;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  photoPath?: string | null;
  isFavorite: boolean;
  createdAt: string;
}